public/download/Applications_educatives/app_second_degre_partie2/requirements-dev.txt
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pyinstaller==6.3.0
black==23.12.1
flake8==7.0.0
